+++ Sources/Relayer/Misc/ConsoleLogging.swift
@@ -1,12 +1,12 @@
 
 import Foundation
 
-public protocol ConsoleLogging {
-    func debug(_ items: Any...)
-
-    func info(_ items: Any...)
-
-    func warn(_ items: Any...)
-
-    func error(_ items: Any...)
-}
+//public protocol ConsoleLogging {
+//    func debug(_ items: Any...)
+//
+//    func info(_ items: Any...)
+//
+//    func warn(_ items: Any...)
+//
+//    func error(_ items: Any...)
+//}
