+++ Sources/WalletConnect/Types/Session/ClientSynchronisation/SessionNotificationParams.swift
@@ -1,5 +1,7 @@
 
 import Foundation
+import WalletConnectUtils
+
 public typealias SessionNotification = SessionType.NotificationParams
 
 extension SessionType {
