+++ Sources/WalletConnectUtils/Logger.swift
@@ -3,7 +3,15 @@
 import Foundation
 
 
-public class ConsoleLogger {
+public protocol ConsoleLogging {
+    func debug(_ items: Any...)
+    func info(_ items: Any...)
+    func warn(_ items: Any...)
+    func error(_ items: Any...)
+}
+
+
+public class ConsoleLogger: ConsoleLogging {
     private var loggingLevel: LoggingLevel
     private var suffix: String
